requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-jose
aiofiles
pytest
httpx
requests
jinja2
websockets